map.js create and adjustment the main.js
